Converted seq to character vector (seq.as.character), started 1letter seq to 3 letter
Problem with converting to 3 letter sequence is that it goes through each possible 1 letter code then converts it - as it gets down the list of 1 letter codes to convert, many have been converted to 3 letter codes and therefore it finds 1 letter codes which are actually 3 letter codes. Perhaps add if(length=1) to each gsub?